# assets/generated/swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cluster Manager VMWare Helper API",
    "version": "v0.1.0",
    "contact": {
      "name": "Michael Venezia",
      "url": "https://gitlab.com/mvenezia/cma-vmware",
      "email": "mvenezia@gmail.com"
    }
  },
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/cluster": {
      "get": {
        "summary": "Will retrieve the status of a cluster and its kubeconfig for connectivity",
        "operationId": "GetCluster",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmavmwareGetClusterReply"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "Name of the cluster to be looked up.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.username",
            "description": "The Username for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.password",
            "description": "The Password for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Cluster"
        ]
      },
      "delete": {
        "summary": "Will delete a cluster",
        "operationId": "DeleteCluster",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmavmwareDeleteClusterReply"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "What is the cluster's name to destroy.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.username",
            "description": "The Username for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.password",
            "description": "The Password for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Cluster"
        ]
      },
      "put": {
        "summary": "Will provision a cluster",
        "operationId": "CreateCluster",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmavmwareCreateClusterReply"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/cmavmwareCreateClusterMsg"
            }
          }
        ],
        "tags": [
          "Cluster"
        ]
      }
    },
    "/api/v1/cluster/list": {
      "get": {
        "summary": "Will retrieve a list of clusters",
        "operationId": "GetClusterList",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmavmwareGetClusterListReply"
            }
          }
        },
        "parameters": [
          {
            "name": "credentials.username",
            "description": "The Username for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credentials.password",
            "description": "The Password for API access.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Cluster"
        ]
      }
    },
    "/api/v1/version": {
      "get": {
        "summary": "Will return version information about api server",
        "operationId": "GetVersionInformation",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/cmavmwareGetVersionReply"
            }
          }
        },
        "tags": [
          "Cluster"
        ]
      }
    }
  },
  "definitions": {
    "GetVersionReplyVersionInformation": {
      "type": "object",
      "properties": {
        "git_version": {
          "type": "string",
          "title": "The tag on the git repository"
        },
        "git_commit": {
          "type": "string",
          "title": "The hash of the git commit"
        },
        "git_tree_state": {
          "type": "string",
          "title": "Whether or not the tree was clean when built"
        },
        "build_date": {
          "type": "string",
          "title": "Date of build"
        },
        "go_version": {
          "type": "string",
          "title": "Version of go used to compile"
        },
        "compiler": {
          "type": "string",
          "title": "Compiler used"
        },
        "platform": {
          "type": "string",
          "title": "Platform it was compiled for / running on"
        }
      }
    },
    "cmavmwareClusterDetailItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "ID of the cluster"
        },
        "name": {
          "type": "string",
          "title": "Name of the cluster"
        },
        "status": {
          "type": "string",
          "title": "What is the status of the cluster"
        },
        "kubeconfig": {
          "type": "string",
          "title": "What is the kubeconfig to connect to the cluster"
        }
      }
    },
    "cmavmwareClusterItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "ID of the cluster"
        },
        "name": {
          "type": "string",
          "title": "Name of the cluster"
        },
        "status": {
          "type": "string",
          "title": "What is the status of the cluster"
        }
      }
    },
    "cmavmwareCreateClusterMsg": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Name of the cluster to be provisioned"
        },
        "provider": {
          "$ref": "#/definitions/cmavmwareCreateClusterProviderSpec",
          "title": "The provider specification"
        }
      }
    },
    "cmavmwareCreateClusterProviderSpec": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "What is the provider - like vmware"
        },
        "k8s_version": {
          "type": "string",
          "title": "The version of Kubernetes"
        },
        "vmware": {
          "$ref": "#/definitions/cmavmwareCreateClusterVMWareSpec",
          "title": "The VMWare specification"
        },
        "high_availability": {
          "type": "boolean",
          "format": "boolean",
          "title": "Whether or not the cluster is HA"
        },
        "network_fabric": {
          "type": "string",
          "title": "The fabric to be used"
        }
      }
    },
    "cmavmwareCreateClusterReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Whether or not the cluster was provisioned by this request"
        },
        "cluster": {
          "$ref": "#/definitions/cmavmwareClusterItem",
          "title": "The details of the cluster request response"
        }
      }
    },
    "cmavmwareCreateClusterVMWareSpec": {
      "type": "object",
      "properties": {
        "credentials": {
          "$ref": "#/definitions/cmavmwareVMWareCredentials",
          "title": "Credentials to build the cluster"
        }
      }
    },
    "cmavmwareDeleteClusterReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Could the cluster be destroyed"
        },
        "status": {
          "type": "string",
          "title": "Status of the request"
        }
      }
    },
    "cmavmwareGetClusterListReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Is the cluster in the system"
        },
        "clusters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/cmavmwareClusterItem"
          },
          "title": "List of clusters"
        }
      }
    },
    "cmavmwareGetClusterReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Is the cluster in the system"
        },
        "cluster": {
          "$ref": "#/definitions/cmavmwareClusterDetailItem"
        }
      }
    },
    "cmavmwareGetVersionReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "If operation was OK"
        },
        "version_information": {
          "$ref": "#/definitions/GetVersionReplyVersionInformation",
          "title": "Version Information"
        }
      },
      "title": "Reply for version request"
    },
    "cmavmwareVMWareCredentials": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "title": "The Username for API access"
        },
        "password": {
          "type": "string",
          "title": "The Password for API access"
        }
      },
      "title": "The credentials to use for creating the cluster"
    }
  },
  "externalDocs": {
    "description": "More about Cluster Manager VMWare Helper API",
    "url": "https://gitlab.com/mvenezia/cma-vmware"
  }
}